data cleaning day 1 completed
In this notebook I was taught how to find how many missing values are in a data set and 2 basic ways on how to fix that problem.  First was to delete the rows or columns containing the missing values.  And the second way was to impute values into the missing values.